requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
openai==1.30.0
faiss-cpu==1.8.0
transformers==4.40.0
torch==2.3.0
peft==0.11.0
streamlit==1.35.0
pypdf==4.2.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
accelerate==0.30.0
bitsandbytes==0.43.1
datasets==2.19.0